backend/requirements.txt
# Django
django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
dj-database-url

# Core
playwright
pandas
openai
python-pptx
Jinja2
python-dotenv
Pillow

# Testing
pytest
pytest-django
pytest-cov
